refactor: extract conversation rendering pipeline
Extract message rendering components and utilities from
CoworkSessionDetail into standalone files for reuse:

- messageDisplayUtils.ts: pure functions and types
- ToolCallGroup.tsx: tool call rendering
- ThinkingBlock.tsx: thinking/reasoning blocks
- AssistantMessageItem.tsx: assistant message + copy
- UserMessageItem.tsx: user message + skill badges
- AssistantTurnBlock.tsx: full turn orchestration
- ConversationTurnsView.tsx: reusable conversation view

CoworkSessionDetail reduced from ~3973 to ~2500 lines.
No behavioral changes — pure refactor.
